requirements.txt
anthropic
launchdarkly-server-sdk
ddtrace
pyairtable
python-dotenv
fastapi
uvicorn[standard]
